* All action area now use a simple Strenght value for their own purposes. *  Unit Strenght and Defense, and Action Area Strenght  are now stored as floats and final heal and damage valued are rounded to int. + Added debug logs for combat math. * Disabled sound manager in scenes as SoundManager script is not working properly now that it was added to SoundManager prefab.

+++ Assets/A_Game/Scripts/MinMinUnit.cs
@@ -24,8 +24,9 @@ public class MinMinUnit : NetworkEntity
         Six
     }
 
-    public int Strength = 1;
-    public int Defense = 1;
+    public float Strength = 1;
+    public float Defense = 1;
+
     public int MaxHealth = 5;
 
     public float EffectScale = 1;
